Not markdown. Post parser working
Backed away from markdown. Too heavyweight for my needs. Now just supports paragraphs and headers along with a title. The first scanner and parser parses a html template, which needs tags showing the title and the body text. This is taken from a txt file formatted very simply. Variables are preceded by '-' e.g. -title || -body. These are delimited with : from the content. The body text is then run througha  simple parser to give it paragraph tags and header tags. This is then inserted into the place in the html template from the first parse pass. It's simple. It works. Next I need to work on generating a number of posts and creating an index of them,