Link support
Added basic link support. Links are delimited first by '+' around the href, and then the link text is allowed before it is delimited with a third '+' from the rest of the text. Currently my parsing is incomplete and links within paragraphs are not found but this should be a simple case to find in the text handling block.