Multiple file support.
Scans a posts directory for .txt files and then writes each of those to a html file in the output folder. Next will be adding an index to each of these files with the title and a link to them in a master index file that should use a different html template.