Index Building
An index of files is now being built. It is hacky, and definitely not very clean, but it works. I should add a support to allow a template snipped be loaded at runtime to give freer access to design of the index link pieces. I should also build in a feature to grab the date from each post's source file so that that stays the same even if it is edited and recompiled in the future.