Static page creation.
Creation of non blog post pages, essentially just pages that don't get indexed by the blog index and live in the home directory. Currently no support for automatically adding those to an index so the links should be hard marked up in the index template or similar. This approach could also essentially work for a generic tag system for the posts with a little more work.