Link in text support
Links in body of paragraphs support. I had run into an issue because of a dangling + character in one of my test posts that threw me off in looking for why it wouldn't compile. I will need to add some error catching to give indications if the program crashes while looking for a closing tag. Also small cleanup of leftover debugging print lines.